+++ src/redturtle/prenotazioni/browser/prenotazioni_context_state.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import itertools
 import json
-import logging
 from datetime import date
 from datetime import datetime
 from datetime import timedelta
@@ -16,6 +15,7 @@ from six.moves import range
 
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni import get_or_create_obj
+from redturtle.prenotazioni import logger
 from redturtle.prenotazioni import tznow
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.adapters.conflict import IConflictManager
@@ -28,8 +28,6 @@ from redturtle.prenotazioni.content.prenotazione_type import PrenotazioneType
 from redturtle.prenotazioni.utilities.dateutils import hm2DT
 from redturtle.prenotazioni.utilities.dateutils import hm2seconds
 from redturtle.prenotazioni.utilities.urls import urlify
-
-logger = logging.getLogger(__name__)
 
 
 class PrenotazioniContextState(BrowserView):